Verificación de correos al crear usuarios completa. Compuesto el error del calculo de la capacidad máxima por salón.

+++ www/public_ftp/iepe/database/migrations/2016_06_17_053350_create_user_activations_table.php
@@ -13,7 +13,7 @@ class CreateUserActivationsTable extends Migration
     public function up()
     {
         Schema::create('user_activations', function (Blueprint $table) {
-            $table->increments('user_id');
+            $table->bigInteger('user_id');
             $table->string('token')->index();
             $table->timestamps();
         });
